+++ src/core/similarity.cpp
@@ -28,7 +28,7 @@ qint64 Similarity::totalPairs(const ExpressionMatrix* emx) const
 {
    EDEBUG_FUNC(this,emx);
 
-   return (qint64) emx->geneSize() * (emx->geneSize() - 1) / 2;
+   return static_cast<qint64>(emx->geneSize()) * (emx->geneSize() - 1) / 2;
 }
 
 
@@ -67,8 +67,8 @@ std::unique_ptr<EAbstractAnalytic::Block> Similarity::makeWork(int index) const
       ELog() << tr("Making work index %1 of %2.\n").arg(index).arg(size());
    }
 
-   qint64 start {index * (qint64) _workBlockSize};
-   qint64 size {min(totalPairs(_input) - start, (qint64) _workBlockSize)};
+   qint64 start {index * static_cast<qint64>(_workBlockSize)};
+   qint64 size {min(totalPairs(_input) - start, static_cast<qint64>(_workBlockSize))};
 
    return unique_ptr<EAbstractAnalytic::Block>(new WorkBlock(index, start, size));
 }
@@ -275,7 +275,7 @@ void Similarity::initialize()
    {
       int numWorkers = max(1, mpi.size() - 1);
 
-      _workBlockSize = min((qint64) 32768, totalPairs(_input) / numWorkers);
+      _workBlockSize = min(32768LL, totalPairs(_input) / numWorkers);
    }
 }
